requirements.txt
asyncio
opencv-python
numpy
vuer
pyrealsense2
pyzed
quaternionic
scipy
pyyaml